% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adaptive_bin_count}
\alias{adaptive_bin_count}
\title{Adaptive Bin Count}
\usage{
adaptive_bin_count(n, default_bins = 5)
}
\arguments{
\item{n}{Sample size}

\item{default_bins}{Maximum bins}
}
\value{
Optimal number of bins
}
\description{
Adaptive Bin Count
}
\keyword{internal}
