% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{AnnotatedGeneLists-class}
\alias{AnnotatedGeneLists-class}
\title{AnnotatedGeneLists class}
\description{
A class to hold a list of GeneList objects, each representing a method.
}
\section{Slots}{

\describe{
\item{\code{inbuilt}}{A list of GeneList objects containing annotations for genes selected with inbuilt, model-specific feature importance.}

\item{\code{permutation}}{A list of GeneList objects containing annotations for genes selected with permutation importance.}
}}

