% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gene_set_stability.R
\name{calculate_gene_set_stability}
\alias{calculate_gene_set_stability}
\title{Calculate Gene Set Stability}
\usage{
calculate_gene_set_stability(selected_features, X_train)
}
\arguments{
\item{selected_features}{A list of selected features. Each element of the list represents a split and should be a named list where the names are the methods and the values are data frames containing the feature importances for that method in that split.}

\item{X_train}{The original training data used for feature selection.}
}
\value{
A list of gene set stability values for each method.
}
\description{
This function calculates the gene set stability for each method across different splits.
}
