% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_overlap_coefficients.R
\name{calculate_overlap_coefficients}
\alias{calculate_overlap_coefficients}
\title{Calculate Overlap and Similarity Coefficients between Feature Lists}
\usage{
calculate_overlap_coefficients(pipeline_results)
}
\arguments{
\item{pipeline_results}{A PipelineResults object containing the fitted pipelines, cross-validation results, selected features,
mean performance, and mean feature importances.}
}
\value{
A list of matrices showing the Overlap, Jaccard, and Soerensen-Dice coefficients for the feature lists for feature importance and permutation importance scores if they are provided
}
\description{
This function calculates the Overlap, Jaccard, and Soerensen-Dice coefficients to quantify
the similarity between feature lists.
}
