% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_scores}
\alias{combine_scores}
\title{Combine Scores into Final Score}
\usage{
combine_scores(pi, u, b, formula = "geometric", weights = c(1, 1, 1))
}
\arguments{
\item{pi}{Stability scores}

\item{u}{Utility scores}

\item{b}{Biology scores}

\item{formula}{Combination formula}

\item{weights}{Weights for (pi, u, b)}
}
\value{
Final combined scores
}
\description{
Combine Scores into Final Score
}
\keyword{internal}
