% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_gene_lists.R
\name{compare_gene_lists}
\alias{compare_gene_lists}
\title{Compare gene lists obtained from the PipelineResults object}
\usage{
compare_gene_lists(input_list, save_plot = FALSE, filename = NULL, ...)
}
\arguments{
\item{input_list}{A PipelineResults object containing the fitted pipelines, cross-validation results, selected features,
mean performance, and mean feature importances.}

\item{save_plot}{A logical value indicating whether to save the heatmap plots to a file or not. Default is FALSE.}

\item{filename}{A character string specifying the filename for the saved heatmap plots (if save_plot = TRUE).}
}
\value{
A grid of heatmaps showing the Overlap, Jaccard, and Soerensen-Dice coefficients for the gene lists.
}
\description{
This function takes a PipelineResults object and creates heatmaps to visualize the overlap between the gene lists obtained
from different feature selection methods. It calculates the Overlap, Jaccard, and Soerensen-Dice coefficients to quantify
the similarity between gene lists.
}
\examples{
# Compare gene lists from the PipelineResults object obtained from fit_and_evaluate_pipelines
data(iris)
X <- iris[,1:4]
y <- iris[,5]
pipeline_results <- fit_and_evaluate_pipelines(X_train = X, y_train = y)
compare_gene_lists(input_list = pipeline_results)
importFrom GeneOverlap modOverlaps
}
