% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_auc}
\alias{compute_auc}
\title{Compute AUC from Predictions}
\usage{
compute_auc(y_true, y_pred)
}
\arguments{
\item{y_true}{True labels}

\item{y_pred}{Predicted probabilities}
}
\value{
AUC score
}
\description{
Compute AUC from Predictions
}
\keyword{internal}
