% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_mi_vectorized}
\alias{compute_mi_vectorized}
\title{Compute Mutual Information (Vectorized)}
\usage{
compute_mi_vectorized(X, y, method = "discrete", n_bins = 5, adaptive = TRUE)
}
\arguments{
\item{X}{Matrix (n_samples x n_genes)}

\item{y}{Binary outcome}

\item{method}{MI estimation method}

\item{n_bins}{Number of bins}

\item{adaptive}{Adaptive binning}
}
\value{
Vector of MI scores (n_genes)
}
\description{
Compute Mutual Information (Vectorized)
}
\keyword{internal}
