% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_mutual_information}
\alias{compute_mutual_information}
\title{Compute Mutual Information}
\usage{
compute_mutual_information(
  x,
  y,
  method = "discrete",
  n_bins = 5,
  adaptive = TRUE
)
}
\arguments{
\item{x}{Numeric vector (gene expression)}

\item{y}{Factor or integer (binary outcome)}

\item{method}{"discrete" or "continuous"}

\item{n_bins}{Number of bins for discrete method}

\item{adaptive}{Use adaptive binning}
}
\value{
Mutual information in nats
}
\description{
Computes mutual information between continuous x and binary y.
}
\examples{
\dontrun{
x <- rnorm(100)
y <- factor(rep(c(0, 1), each = 50))
mi <- compute_mutual_information(x, y)
}

}
