% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_cv_folds}
\alias{create_cv_folds}
\title{Create Cross-Validation Folds}
\usage{
create_cv_folds(y, K = 5, R = 20, random_seed = 123)
}
\arguments{
\item{y}{Factor, outcome variable}

\item{K}{Number of folds}

\item{R}{Number of repeats}

\item{random_seed}{Random seed}
}
\value{
List of train/test indices
}
\description{
Creates stratified K-fold CV with R repeats.
}
\examples{
\dontrun{
y <- factor(rep(c("A", "B"), each = 50))
folds <- create_cv_folds(y, K = 5, R = 10)
}

}
