% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_pipelines}
\alias{create_pipelines}
\title{Create Pipelines}
\usage{
create_pipelines(
  feature_selection_methods,
  preprocessing_steps,
  selected_methods,
  classifier,
  fs_param_grids
)
}
\arguments{
\item{feature_selection_methods}{A list of feature selection methods to use for the pipelines.}

\item{preprocessing_steps}{A list of preprocessing steps to use for the pipelines.}

\item{selected_methods}{A vector of names of feature selection methods to use from the default set.}

\item{classifier}{A Scikit-learn classifier to use as the final step in the pipelines.}

\item{fs_param_grids}{A list of hyperparameter grids for the feature selection methods.}
}
\value{
A list of Scikit-learn pipelines.
}
\description{
This function creates a list of Scikit-learn pipelines using the specified feature selection methods, preprocessing steps, and classifier.
}
