% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{create_similarity_cache}
\alias{create_similarity_cache}
\title{Create Similarity Cache Environment}
\usage{
create_similarity_cache()
}
\value{
Environment for storing GO similarity scores
}
\description{
Create Similarity Cache Environment
}
\keyword{internal}
