% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSelectR.R
\name{create_test_metrics_df}
\alias{create_test_metrics_df}
\title{Create a Dataframe of Test Metrics}
\usage{
create_test_metrics_df(test_metrics)
}
\arguments{
\item{test_metrics}{A list or dataframe of test metrics.}
}
\value{
A dataframe with the processed test metrics.
}
\description{
Create a Dataframe of Test Metrics
}
