% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_results.R
\name{custom_compute_SS_distances}
\alias{custom_compute_SS_distances}
\title{Visualize Results of ViSEAGO Analysis}
\usage{
custom_compute_SS_distances(GO_SS, distance)
}
\arguments{
\item{results}{A data frame of enriched GO terms.}

\item{distance_GO_SS}{The distance metric for semantic similarity analysis of GO terms (default: "Wang").}

\item{distance_GO_cluster_SS}{The distance metric for semantic similarity analysis of GO clusters (default: "BMA").}

\item{tree_distance}{The distance metric for hierarchical clustering of GO terms and clusters (default: "Wang").}

\item{tree_aggreg_method}{The agglomeration method for hierarchical clustering of GO terms and clusters (default: "ward.D2").}

\item{heatmap_dynamic_deepSplit}{The deepSplit argument for dynamicTreeCut::cutreeDynamic (default: 2).}

\item{heatmap_dynamic_minClusterSize}{The minimum cluster size for dynamicTreeCut::cutreeDynamic (default: 10).}

\item{compute_SS_distances_args}{A list of additional arguments to be passed to custom_compute_SS_distances().}

\item{GOterms_heatmap_args}{A list of additional arguments to be passed to custom_GOterms_heatmap().}

\item{GOclusters_heatmap_args}{A list of additional arguments to be passed to custom_GOclusters_heatmap().}
}
\value{
A list containing two elements: "GO_terms_heatmap" and "GO_clusters_heatmap", which are objects of class 'GOTermsHeatmap' and 'GOClustersHeatmap' respectively.
}
\description{
This function creates and returns heatmaps of Gene Ontology (GO) terms and clusters using the ViSEAGO package.
}
\examples{
\dontrun{
library(GeneSelectR)
result_list <- visualize_results(results)
}

}
