% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSelectR.R
\name{define_sklearn_modules}
\alias{define_sklearn_modules}
\title{Define Python modules and scikit-learn submodules}
\usage{
define_sklearn_modules()
}
\value{
A list containing the definitions for the Python modules and submodules.
}
\description{
Define Python modules and scikit-learn submodules
}
