% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{discretize_continuous}
\alias{discretize_continuous}
\title{Discretize Continuous Values}
\usage{
discretize_continuous(x, n_bins = 5, adaptive = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{n_bins}{Number of bins}

\item{adaptive}{Use adaptive binning}
}
\value{
Integer vector of bin assignments
}
\description{
Discretize Continuous Values
}
\keyword{internal}
