% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_heatmap.R
\name{display_heatmap}
\alias{display_heatmap}
\title{Display a heatmap for GO terms or GO clusters
old ViSEAGO asset}
\usage{
display_heatmap(
  heatmap_object,
  heatmap_type,
  save_plot = FALSE,
  output_file = NULL
)
}
\arguments{
\item{heatmap_object}{An object of class "GOTermsHeatmap" or "GOClustersHeatmap" generated by the visualize_results function.}

\item{heatmap_type}{A character string specifying the type of heatmap to display. Must be either "GOterms" or "GOclusters".}

\item{save_plot}{A logical value indicating whether to save the heatmap as a PNG image. If set to FALSE (default), the interactive heatmap is displayed.}

\item{output_file}{A character string specifying the file path for the output PNG image when save_plot is set to TRUE. If save_plot is FALSE, this parameter is ignored.}
}
\value{
If save_plot is FALSE, displays the interactive heatmap. If save_plot is TRUE, saves the heatmap as a static PNG image at the specified output_file path.
}
\description{
Display a heatmap for GO terms or GO clusters
old ViSEAGO asset
}
\examples{
\dontrun{
results <- # Load results from a previous step
visualized_results <- visualize_results(results)
display_heatmap(visualized_results$GO_terms_heatmap, "GOterms")
display_heatmap(visualized_results$GO_clusters_heatmap, "GOclusters", save_plot = TRUE, output_file = "GOclusters_heatmap.png")
}
}
\keyword{hidden}
\keyword{internal}
