% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{fit_and_evaluate_pipelines}
\alias{fit_and_evaluate_pipelines}
\title{Fit and Evaluate Pipelines with Hyperparameter Tuning}
\usage{
fit_and_evaluate_pipelines(
  X_train,
  y_train,
  pipelines = NULL,
  feature_selection_methods = NULL,
  selected_methods = NULL,
  njobs = -1L,
  n_splits = 2L
)
}
\arguments{
\item{X_train}{A matrix or data frame of training data with features as columns and observations as rows.}

\item{y_train}{A vector of training labels corresponding to the rows of X_train.}

\item{pipelines}{An optional list of pre-defined pipelines to use for fitting and evaluation. If this argument is provided, the feature selection methods and preprocessing steps will be ignored.}

\item{feature_selection_methods}{An optional list of feature selection methods to use for fitting and evaluation. If this argument is not provided, a default set of feature selection methods will be used.}

\item{selected_methods}{An optional vector of names of feature selection methods to use from the default set. If this argument is provided, only the specified methods will be used.}
}
\value{
A list with the following elements:
\item{fitted_pipelines}{A list of the fitted pipelines.}
\item{cv_results}{A list of the cross-validation results for each pipeline.}
}
\description{
This function fits a set of pipelines with hyperparameter tuning on a given training set and evaluates their performance using cross-validation.
}
\examples{
# Fit and evaluate pipelines using the default feature selection methods
data(iris)
X <- iris[,1:4]
y <- iris[,5]
results <- fit_and_evaluate_pipelines(X_train = X, y_train = y)

# Fit and evaluate pipelines using a subset of the default feature selection methods
results <- fit_and_evaluate_pipelines(X_train = X, y_train = y, selected_methods = c("Univariate", "RFE"))

# Fit and evaluate pipelines using user-defined feature selection methods
fs_methods <- list("Lasso" = select_model(lasso(penalty = 'l1', C = 0.1, solver = 'saga'), threshold = 'median'))
results <- fit_and_evaluate_pipelines(X_train = X, y_train = y, feature_selection_methods = fs_methods)

}
