% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{GeneList-class}
\alias{GeneList-class}
\title{GeneList class}
\description{
A class to hold annotated gene list for a single method.
}
\section{Slots}{

\describe{
\item{\code{SYMBOL}}{A character vector of gene names.}

\item{\code{ENSEMBL}}{A character vector of Ensembl IDs.}

\item{\code{ENTREZID}}{A character vector of Entrez IDs.}
}}

