% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_GO_heatmaps.R
\name{generate_GO_heatmaps}
\alias{generate_GO_heatmaps}
\title{Generate Heatmaps for Semantic Similarity Analysis
old ViSEAGO asset
This function creates and returns heatmaps of Gene Ontology (GO) terms and clusters using the ViSEAGO package.}
\usage{
generate_GO_heatmaps(
  GO_SS_and_clusters,
  tree_distance = "Wang",
  tree_aggreg_method = "ward.D2",
  heatmap_dynamic_deepSplit = 2,
  heatmap_dynamic_minClusterSize = 10,
  distance_GO_cluster_SS = "BMA",
  GO_terms_distance = "BMA"
)
}
\arguments{
\item{GO_SS_and_clusters}{A list containing "GO_SS" and "GO_cluster_SS", which are objects containing semantic similarity distances.}

\item{tree_distance}{The distance metric for hierarchical clustering of GO terms and clusters (default: "Wang").}

\item{tree_aggreg_method}{The agglomeration method for hierarchical clustering of GO terms and clusters (default: "ward.D2").}

\item{heatmap_dynamic_deepSplit}{The deepSplit argument for dynamicTreeCut::cutreeDynamic (default: 2).}

\item{heatmap_dynamic_minClusterSize}{The minimum cluster size for dynamicTreeCut::cutreeDynamic (default: 10).}
}
\value{
A list containing two elements: "GO_terms_heatmap" and "GO_clusters_heatmap", which are objects of class 'GOTermsHeatmap' and 'GOClustersHeatmap' respectively.
}
\description{
Generate Heatmaps for Semantic Similarity Analysis
old ViSEAGO asset
This function creates and returns heatmaps of Gene Ontology (GO) terms and clusters using the ViSEAGO package.
}
\examples{
\dontrun{
results_heatmap <- generate_GO_heatmaps(GO_SS_and_clusters)
}
}
\references{
To use ViSEAGO in published research, please cite:
\link{citation}
}
\keyword{hidden}
\keyword{internal}
