% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_overlap_heatmaps.R
\name{generate_overlap_heatmaps}
\alias{generate_overlap_heatmaps}
\title{Generate Heatmaps to Visualize Overlap and Similarity Coefficients between Feature Lists}
\usage{
generate_overlap_heatmaps(coefficients, save_plot = FALSE, filename = NULL)
}
\arguments{
\item{coefficients}{A list of matrices showing the Overlap, Jaccard, and Soerensen-Dice coefficients for the feature lists.}

\item{save_plot}{A logical value indicating whether to save the heatmap plots to a file or not. Default is FALSE.}

\item{filename}{A character string specifying the filename for the saved heatmap plots (if save_plot = TRUE).}
}
\value{
A grid of heatmaps showing the Overlap, Jaccard, and Soerensen-Dice coefficients for the feature lists.
}
\description{
This function takes a list of matrices of overlap and similarity coefficients and generates heatmaps to visualize them.
}
