% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSelectR-package.R
\docType{package}
\name{GeneSelectR-package}
\alias{GeneSelectR}
\alias{GeneSelectR-package}
\title{GeneSelectR: predictive gene selection and biological interpretation}
\description{
GeneSelectR estimates gene-selection recurrence and excluded-sample
predictive contribution across repeated elastic-net models. Outcome
permutations provide gene-specific reference values. Gene Ontology,
Hallmark gene sets, Open Targets, and STRING support downstream biological
interpretation.
}
\seealso{
Useful links:
\itemize{
    \item \url{https://github.com/dzhakparov/GeneSelectR-2.0}
    \item Report bugs at
    \url{https://github.com/dzhakparov/GeneSelectR-2.0/issues}
}

}
\author{
\strong{Maintainer}: Damir Zhakparov \email{dzhakparov@gmail.com}
(\href{https://orcid.org/0000-0001-7175-0843}{ORCID})

Authors:
\itemize{
    \item Damir Zhakparov \email{dzhakparov@gmail.com}
    (\href{https://orcid.org/0000-0001-7175-0843}{ORCID})
}

}
\keyword{internal}
