% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Locate the package cache directory}
\usage{
get_cache_dir()
}
\value{
Character path to the cache directory.
}
\description{
Uses \code{tools::R_user_dir()} to obtain a platform-specific cache location.
}
\keyword{internal}
