% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChildNodes.R
\name{get_child_nodes}
\alias{get_child_nodes}
\title{Get Child Nodes From Cluster Summaries}
\usage{
get_child_nodes(go_ids, cluster_summaries)
}
\arguments{
\item{go_ids}{A character vector of GO IDs to retrieve child nodes for.}

\item{cluster_summaries}{A named list of cluster summary data frames.}
}
\value{
A named list of data frames, where each data frame contains the child nodes of a given GO ID that appear in a cluster's summary, and a list of the fraction of the child terms for each GO ID that were found in the cluster summaries.
}
\description{
This function retrieves child nodes for given parent nodes (GO IDs) and filters them according to given cluster summaries. It also calculates the fraction of the child terms retrieved from the specified go ID in the cluster summaries.
}
\examples{
\dontrun{
go_ids <- c('GO:0002376', 'GO:0044419')
cluster_summaries <- list(
  lasso = lasso.cluster_summary,
  rfe_lr = rfe_lr.cluster_summary,
  rfe_rf = rfe_rf.cluster_summary,
  uni = uni.cluster_summary,
  dge = dge.cluster_summary
)
results <- get_child_nodes(go_ids, cluster_summaries)
}
}
