% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_enrichment_analysis.R
\name{GO_enrichment_analysis}
\alias{GO_enrichment_analysis}
\title{Perform gene set enrichment analysis using ViSEAGO and topGO}
\usage{
GO_enrichment_analysis(
  gene_sets,
  background,
  organism = "org.Hs.eg.db",
  ont = "BP",
  nodeSize = 5,
  algorithm = "elim",
  statistic = "fisher",
  cutOff = 0.01,
  mergeCutOff = 0.05
)
}
\arguments{
\item{gene_sets}{A list of named character vectors containing gene sets.}

\item{background}{A character vector representing the background gene set.}

\item{organism}{A character string corresponding to the organism of interest. Default: "org.Hs.eg.db" (for human).}

\item{ont}{A character string representing GO term ontology. Default: "BP" (for Biological Process).}

\item{nodeSize}{An integer representing the minimum number of annotated genes for a GO term. Default: 5.}

\item{algorithm}{A character string specifying the algorithm to use in topGO. Default: "elim".}

\item{statistic}{A character string specifying the statistical test to use in topGO. Default: "fisher".}

\item{cutOff}{A numeric value specifying the significance cutoff in topGO. Default: 0.01.}

\item{mergeCutOff}{A numeric value specifying the significance cutoff for merging results in ViSEAGO. Default: 0.05.}
}
\value{
A data frame containing the merged enrichment results for all gene sets.
}
\description{
This function performs gene set enrichment analysis using ViSEAGO and topGO.
The functionality is largely based on methods from the ViSEAGO package by the following authors:
Amélie L. F. Gaudin, Marie Chabaud, Bruno Spataro, Christophe Klopp, Thibaut Jombart, Stéphanie Sidibé-Bocs.
}
\examples{
\dontrun{
# Assuming all.genes and gene_sets are defined
background = as.character(all.genes$entrezid)
BP_sResults <- perform_GO_enrichment_analysis(gene_sets, background, organism = "org.Hs.eg.db")
}
}
\references{
To use ViSEAGO in published research, please cite:
\link{citation}
}
