% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_env_setup.R
\name{install_python_packages}
\alias{install_python_packages}
\title{Install necessary Python packages in a specific Conda environment}
\usage{
install_python_packages(conda_env = "GeneSelectR_env")
}
\arguments{
\item{conda_env}{The name of the Conda environment to use.}
}
\value{
Installs necessary version of Python packages into the GeneSelectR_env.
}
\description{
This function installs the necessary Python packages in a specific Conda environment.
}
