% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{load_python_packages}
\alias{load_python_packages}
\title{Load Python Modules}
\usage{
load_python_packages()
}
\value{
If successful, the global variables for each Python module are set.
Otherwise, it will return an error message.
}
\description{
This function imports necessary Python modules with delayed loading.
}
\details{
Imports the following Python modules with \code{delay_load} set to TRUE:
\itemize{
\item sklearn
\item pandas
\item numpy
\item boruta
\item sys
\item multiprocessing
\item skopt
}
}
\keyword{internal}
