% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{onLoad}
\alias{onLoad}
\alias{load_python_packages}
\title{Load Python Modules}
\usage{
load_python_packages()
}
\arguments{
\item{libname}{The name of the library.}

\item{pkgname}{The name of the package.}
}
\value{
If not in interactive mode, the function will return NULL. Otherwise, it will import the Python modules and return a message indicating that the import was successful.
}
\description{
This function is called when the package is loaded. It checks if Conda is installed and imports necessary Python modules.
}
\details{
If Conda is not installed, the function will stop and prompt the user to install Conda. If Conda is installed, the function will import the Python modules with delay_load set to TRUE.
}
\keyword{internal}
