% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{percentile01}
\alias{percentile01}
\title{Percentile Normalization}
\usage{
percentile01(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Normalized vector in (0,1)
}
\description{
Normalizes values to (0,1) range using percentile ranks, handling zeros.
}
\examples{
x <- c(0, 1, 2, 3, 4, 5)
percentile01(x)

}
