% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_GO_enrichment.R
\name{perform_GO_enrichment_analysis}
\alias{perform_GO_enrichment_analysis}
\title{Perform gene set enrichment analysis using ViSEAGO and topGO}
\usage{
perform_GO_enrichment_analysis(gene_sets, background)
}
\arguments{
\item{background}{A character vector representing the background gene set.}

\item{pipeline_results}{A PipelineResults object containing a named list of data frames
with a "features" column as gene sets.}

\item{myGENE2GO}{A data frame containing GO annotations from Bioconductor.}

\item{additional_gene_sets}{(optional) A named list of character vectors containing
additional user-defined gene sets.}
}
\value{
A data frame containing the merged enrichment results for all gene sets.
}
\description{
This function takes a PipelineResults object containing selected_features
as a named list of data frames with a "features" column, a background gene set,
GO annotations, and optionally additional gene sets provided by the user.
It performs gene set enrichment analysis for each gene set and returns the merged results.
}
\examples{
# Assuming all.genes, PipelineResults, and myGENE2GO are defined
background = as.character(all.genes$entrezid)
BP_sResults <- perform_analysis(PipelineResults, background, myGENE2GO)


}
