% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSelectR.R
\name{perform_grid_search}
\alias{perform_grid_search}
\title{Perform Grid Search or Random Search for Hyperparameter Tuning}
\usage{
perform_grid_search(
  X_train,
  y_train,
  pipeline,
  scoring,
  params,
  search_type,
  n_iter,
  njobs,
  modules,
  random_state
)
}
\arguments{
\item{X_train}{Training data for predictors.}

\item{y_train}{Training data for outcomes.}

\item{pipeline}{A pipeline specifying the steps for feature selection and model training.}

\item{scoring}{A string representing what scoring metric to use for hyperparameter adjustment. Default value is 'accuracy'}

\item{params}{A list of parameters or parameter distributions to search over.}

\item{search_type}{A character string specifying the type of search ('grid' or 'random').}

\item{n_iter}{The number of parameter settings that are sampled in a random search.}

\item{njobs}{The number of CPU cores to use.}

\item{modules}{A list containing the definitions for the Python modules and submodules.}

\item{random_state}{An integer value setting the random seed for feature selection algorithms and cross validation procedure. By default set to NULL to use different random seed every time an algorithm is used. For reproducibility could be fixed, otherwise for an unbiased estimation should be left as NULL.}
}
\value{
An object of the optimal model found during the search.
}
\description{
Perform Grid Search or Random Search for Hyperparameter Tuning
}
