% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{PipelineResults-class}
\alias{PipelineResults-class}
\title{PipelineResults class}
\description{
A class to hold the results of fitting and evaluating pipelines.
}
\section{Slots}{

\describe{
\item{\code{fitted_pipelines}}{A list of the fitted pipelines.}

\item{\code{cv_results}}{A list of the cross-validation results for each pipeline.}

\item{\code{selected_features}}{A list of the selected features for each pipeline.}

\item{\code{mean_performance}}{A data frame of the mean performance for each feature selection method across all splits.}

\item{\code{mean_feature_importances}}{A list of the mean feature importances for each method across all splits.}
}}

