% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_upset.R
\name{plot_upset}
\alias{plot_upset}
\title{Plot Feature Overlaps Using UpSet Plots}
\usage{
plot_upset(pipeline_results, custom_lists = NULL)
}
\arguments{
\item{pipeline_results}{A PipelineResults object containing the fitted pipelines, cross-validation results, selected features,
mean performance, and mean feature importances.}

\item{custom_lists}{An optional named list of character vectors. Each character vector should contain feature names.
The names of the list will be used as names in the UpSet plots.}
}
\value{
A named list containing two UpSet plots accessible as \code{result$plot1} for feature importances
and \code{result$plot2} for permutation importances.
}
\description{
This function produces separate UpSet plots for inbuilt feature importances and permutation importances,
allowing you to visualize the overlap of feature lists. Optionally, you can include custom lists.
}
\examples{
\dontrun{
pipeline_results <- PipelineResults$new(...)
custom_lists <- list("custom1" = c("f1", "f2"), "custom2" = c("f3", "f4"))
result <- plot_feature_overlap_upset(pipeline_results, custom_lists)
print(result$plot1)
print(result$plot2)
}

}
