% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSelectR2.R
\name{print.geneselectr2_result}
\alias{print.geneselectr2_result}
\title{Print a GeneSelectR 2.0 result}
\usage{
\method{print}{geneselectr2_result}(x, ...)
}
\arguments{
\item{x}{A \code{geneselectr2_result} object.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a GeneSelectR 2.0 result
}
