% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSelectR.R
\name{print.geneselectr_result}
\alias{print.geneselectr_result}
\title{Print a GeneSelectR result}
\usage{
\method{print}{geneselectr_result}(x, ...)
}
\arguments{
\item{x}{A \code{geneselectr_result} object.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a GeneSelectR result
}
