% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{python-modules}
\alias{python-modules}
\alias{sklearn}
\title{Global references to Python modules}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 1.
}
\usage{
sklearn
}
\description{
These are global references to Python modules that will be initialized in .onLoad.
}
\details{
The following Python modules are referenced:
\itemize{
\item sklearn
\item pandas
\item numpy
\item lightgbm
\item xgboost
\item boruta
\item sys
\item multiprocessing
}
}
\keyword{internal}
