% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifyenrichment_test.R
\name{run_simplify_enrichment}
\alias{run_simplify_enrichment}
\title{Run simplifyGOFromMultipleLists with specified measure and method}
\usage{
run_simplify_enrichment(
  fs_GO_results,
  padj_column,
  padj_cutoff,
  ont,
  measure,
  method,
  ...
)
}
\arguments{
\item{fs_GO_results}{A list of dataframes containing GO enrichment results for feature selection methods. The GO object for the simplifyGOFromMultipleLists function.}

\item{padj_column}{Character. The column name for the p-value adjustment.}

\item{padj_cutoff}{Numeric. The cutoff for the p-value adjustment.}

\item{ont}{Character. The ontology for the simplifyGOFromMultipleLists function.}

\item{measure}{Character. The semantic similarity measure for the simplifyGOFromMultipleLists function.}

\item{method}{Character. The clustering method for the simplifyGOFromMultipleLists function.}

\item{...}{Other parameters that can be passed to simplifyGOFromMultipleLists}
}
\value{
The result of the simplifyGOFromMultipleLists function.
}
\description{
This function is simply a wrapper for the simplifyGOFromMultipleLists function in the simplifyEnrichment package,
created for the ease of data input. All credit for the underlying functionality goes to the authors
of the simplifyEnrichment package.
}
