% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_similarity_analysis.R
\name{semantic_similarity_analysis}
\alias{semantic_similarity_analysis}
\title{Calculate Semantic Similarity Distances
Old ViSEAGO asset
This function calculates semantic similarity distances between Gene Ontology (GO) terms using the ViSEAGO package.}
\usage{
semantic_similarity_analysis(
  enriched_results,
  distance_GO_SS = "Wang",
  distance_GO_cluster_SS = "Wang",
  organism = "org.Hs.eg.db"
)
}
\arguments{
\item{enriched_results}{A data frame of enriched GO terms.}

\item{distance_GO_SS}{The distance metric for semantic similarity analysis of GO terms (default: "Wang").}

\item{distance_GO_cluster_SS}{The distance metric for semantic similarity analysis of GO clusters (default: "BMA").}

\item{organism}{A character string corresponding to the organism of interest. Default: "org.Hs.eg.db" (for human).}
}
\value{
A list containing two elements: "GO_SS" and "GO_cluster_SS", which are objects containing semantic similarity distances.
}
\description{
Calculate Semantic Similarity Distances
Old ViSEAGO asset
This function calculates semantic similarity distances between Gene Ontology (GO) terms using the ViSEAGO package.
}
\examples{
\dontrun{
GO_SS_and_clusters <- compute_GO_SS_distances(enriched_results)
}
}
\references{
To use ViSEAGO in published research, please cite:
\link{citation}
}
\keyword{hidden}
