% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{sim_jiang}
\alias{sim_jiang}
\title{Jiang-Conrath Similarity}
\usage{
sim_jiang(mica_ic, ic1, ic2)
}
\description{
distance = IC(t1) + IC(t2) - 2 * IC(MICA)
sim = 1 / (1 + distance)
Range: (0, 1]. Emphasizes differences more strongly.
}
\keyword{internal}
