% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{sim_lin}
\alias{sim_lin}
\title{Lin Similarity}
\usage{
sim_lin(mica_ic, ic1, ic2)
}
\description{
sim = 2 * IC(MICA) / (IC(t1) + IC(t2))
Range: \link{0, 1}. Symmetric normalization. Most widely used.
}
\keyword{internal}
