% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{sim_resnik}
\alias{sim_resnik}
\title{Resnik Similarity (max-normalized)}
\usage{
sim_resnik(mica_ic, ic1, ic2)
}
\description{
sim = IC(MICA) / max(IC(t1), IC(t2))
Range: \link{0, 1}. Simple, robust.
}
\keyword{internal}
