% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{steps_to_tuples}
\alias{steps_to_tuples}
\title{Convert Steps to Tuples}
\usage{
steps_to_tuples(steps)
}
\arguments{
\item{steps}{A list of steps to convert to tuples.}
}
\value{
A list of tuples, where each tuple represents a step in a Scikit-learn pipeline.
The tuple contains the name of the step and the corresponding step object.
}
\description{
This function converts a list of steps to tuples for use in a Scikit-learn pipeline.
}
