% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{TestMetrics-class}
\alias{TestMetrics-class}
\title{Class Union for Test metrics output that could contain either a dataframe or a lists}
\description{
A class union that can contain either a data frame or a list.
}
\seealso{
\link{setClassUnion}
}
\keyword{internal}
