% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UrbanRandomSubset.R
\docType{data}
\name{UrbanRandomSubset}
\alias{UrbanRandomSubset}
\title{SOS-ALL dataset Urban subset}
\format{
A data frame with 60 rows and 1001 columns:
\describe{
\item{var1}{Description of var1}
\item{var2}{Description of var2}
...
}
}
\usage{
UrbanRandomSubset
}
\description{
A dataset containing 1000 randomly selected columns from the complete SOS-ALL
dataset.
}
\keyword{datasets}
