% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_results.R
\name{visualize_results}
\alias{visualize_results}
\title{Visualize Results of ViSEAGO Analysis
old ViSEAGO asset
This function creates and returns heatmaps of Gene Ontology (GO) terms and clusters using the ViSEAGO package.}
\usage{
visualize_results(
  results,
  distance_GO_SS = "Wang",
  distance_GO_cluster_SS = "BMA",
  tree_distance = "Wang",
  tree_aggreg_method = "ward.D2",
  heatmap_dynamic_deepSplit = 2,
  heatmap_dynamic_minClusterSize = 10
)
}
\arguments{
\item{results}{A data frame of enriched GO terms.}

\item{distance_GO_SS}{The distance metric for semantic similarity analysis of GO terms (default: "Wang").}

\item{distance_GO_cluster_SS}{The distance metric for semantic similarity analysis of GO clusters (default: "BMA").}

\item{tree_distance}{The distance metric for hierarchical clustering of GO terms and clusters (default: "Wang").}

\item{tree_aggreg_method}{The agglomeration method for hierarchical clustering of GO terms and clusters (default: "ward.D2").}

\item{heatmap_dynamic_deepSplit}{The deepSplit argument for dynamicTreeCut::cutreeDynamic (default: 2).}

\item{heatmap_dynamic_minClusterSize}{The minimum cluster size for dynamicTreeCut::cutreeDynamic (default: 10).}
}
\value{
A list containing two elements: "GO_terms_heatmap" and "GO_clusters_heatmap", which are objects of class 'GOTermsHeatmap' and 'GOClustersHeatmap' respectively.
}
\description{
Visualize Results of ViSEAGO Analysis
old ViSEAGO asset
This function creates and returns heatmaps of Gene Ontology (GO) terms and clusters using the ViSEAGO package.
}
\examples{
\dontrun{
library(GeneSelectR)
result_list <- visualize_results(results)
}
}
\references{
See ViSEAGO for more information on the underlying functions used:
Lemoine F, Labedan B, Froidevaux C (2018). “ViSEAGO: a Bioconductor package for clustering biological functions using Gene Ontology and semantic similarity.” BioData Mining, 11, 24. doi: 10.1186/s13040-018-0181-9.
}
\keyword{hidden}
