% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{AnnotatedGeneLists-class}
\alias{AnnotatedGeneLists-class}
\title{AnootatedGeneLists class}
\description{
A class to hold a list of GeneList objects, each representing a method.
}
\section{Slots}{

\describe{
\item{\code{gene_lists}}{A list of GeneList objects.}
}}

