% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_entrezid.R
\name{convert_to_entrezid}
\alias{convert_to_entrezid}
\title{Convert to ENTREZIDs}
\usage{
convert_to_entrezid(pipeline_results, custom_lists = NULL)
}
\arguments{
\item{pipeline_results}{A PipelineResults object containing a named list of data frames.}

\item{custom_lists}{(optional) A named list of character vectors containing additional user-defined gene sets.}
}
\value{
A list of gene sets converted to ENTREZIDs.
}
\description{
This function converts a list of gene lists to ENTREZIDs.
}
