% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_env_setup.R
\name{create_conda_env}
\alias{create_conda_env}
\title{Create a specific Conda environment}
\usage{
create_conda_env(conda_env = "GeneSelectR_env")
}
\arguments{
\item{conda_env}{The name of the Conda environment to create.}
}
\value{
Creates conda environment for GeneSelectR package called 'GeneSelectR_env'. If environment already exists, returns a message indicating that the environment is there.
}
\description{
This function creates a Conda environment if it doesn't already exist.
}
\keyword{hidden}
