% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_setup.R
\name{.onLoad}
\alias{.onLoad}
\title{Load the package and set up the Conda environment}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{The name of the package.}

\item{pkgname}{The path to the library where the package is installed.}
}
\description{
This function is automatically run when the package is loaded. It checks if Conda is installed, creates a Conda environment if it doesn't exist, sets the R session to use the Conda environment if it's not already using it, and imports the Python libraries.
}
\details{
The name of the Conda environment is obtained from the "FitSelect.conda_env" option or the "FITSELECT_CONDA_ENV" environment variable. If neither of these is set, a default name of "FitSelectEnv" is used. The Conda environment is created if it doesn't already exist. The R session is set to use the Conda environment if the "RETICULATE_PYTHON" environment variable is not set to the Python interpreter in the Conda environment. The Python libraries "sklearn", "pandas", "numpy", "lightgbm", "xgboost", and "boruta" are imported.
}
