% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_feature_importances}
\alias{get_feature_importances}
\title{Get Feature Importances}
\usage{
get_feature_importances(pipeline, X_train, pipeline_name, iter)
}
\arguments{
\item{pipeline}{A Scikit-learn pipeline object with a Gradient Boosting Classifier
as the final step.}

\item{X_train}{A DataFrame containing the training data.}
}
\value{
A list containing the selected feature names and their importances, or NULL
if the classifier is not a Gradient Boosting Classifier or the feature selector
doesn't have the 'get_support' method.
}
\description{
This function extracts feature importances from a Scikit-learn pipeline
that has a Gradient Boosting Classifier as the final step.
}
\examples{
\dontrun{
# Assuming you have a Scikit-learn pipeline 'my_pipeline' and training data 'X_train'
feature_importances <- get_feature_importances(my_pipeline, X_train)
# Extract the selected feature names and their importances
selected_features <- feature_importances$selected_features
importances <- feature_importances$importances
}
}
