% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_enrichment_analysis.R
\name{GO_enrichment_analysis}
\alias{GO_enrichment_analysis}
\title{Perform gene set enrichment analysis using clusterProfiler}
\usage{
GO_enrichment_analysis(
  annotated_gene_lists,
  list_type = "inbuilt",
  background,
  organism = "org.Hs.eg.db",
  keyType = "ENTREZID",
  ont = "BP",
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.2,
  pAdjMethod = "fdr"
)
}
\arguments{
\item{annotated_gene_lists}{An AnnotatedGeneLists object containing a list of GeneList objects.}

\item{list_type}{A type of AnnotatedGeneList from annotate_gene_lists function. Either 'inbuilt' or 'permutation'. (default: 'inbuilt')}

\item{background}{A character vector representing the background gene set.}

\item{organism}{A character string corresponding to the organism of interest. Default: "org.Hs.eg.db" (for human).}

\item{keyType}{A character string indicating the type of gene identifiers. Default: "ENTREZID".}

\item{ont}{A character string representing GO term ontology. Default: "BP" (for Biological Process).}

\item{pvalueCutoff}{A numeric value specifying the significance cutoff. Default: 0.05.}

\item{qvalueCutoff}{A numeric value specifying the q-value cutoff. Default: 0.2.}

\item{pAdjMethod}{A p-value adjustment method. Should be the one from "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". Default is 'fdr}
}
\value{
A list containing the enrichment results for all gene sets (excluding "background").
}
\description{
This function performs gene set enrichment analysis on a list of gene sets extracted from an AnnotatedGeneLists object.
}
\examples{
\dontrun{
# Assuming annotated_gene_lists and background are defined
results <- GO_enrichment_analysis(annotated_gene_lists, background, organism = "org.Hs.eg.db", keyType = "ENTREZID")
}
}
\references{
To use clusterProfiler in published research, please cite:
Yu G, Wang LG, Han Y, He QY. clusterProfiler: an R package for comparing biological themes among gene clusters. OMICS: A Journal of Integrative Biology. 2012;16(5):284-287. doi:10.1089/omi.2011.0118.
}
