% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{onLoad}
\alias{onLoad}
\alias{.onLoad}
\title{Package Load Function}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{The name of the library.}

\item{pkgname}{The name of the package.}
}
\value{
If not in interactive mode, the function will return NULL. Otherwise, it returns a message indicating that the import was successful.
}
\description{
This function is called when the package is loaded. It checks if Conda is installed and calls \code{load_python_packages} to import Python modules.
}
\details{
If Conda is not installed, the function will stop and prompt the user to install Conda. If Conda is installed, it will proceed to import Python modules.
}
\keyword{internal}
