% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{percentile01}
\alias{percentile01}
\title{Convert positive values to percentile scores}
\usage{
percentile01(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector between zero and one.
}
\description{
Zero values remain zero. Positive values are assigned their fractional rank.
}
\keyword{internal}
