% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{PipelineResults-class}
\alias{PipelineResults-class}
\title{PipelineResults class}
\description{
A class to hold the results of fitting and evaluating pipelines.
}
\section{Slots}{

\describe{
\item{\code{fitted_pipelines}}{A list of the fitted pipelines.}

\item{\code{cv_results}}{A list of the cross-validation results for each pipeline.}

\item{\code{inbuilt_feature_importance}}{A list of the inbuilt mean feature importances for each method across all splits.}

\item{\code{permutation_importance}}{A list of the permutation importances for each method.}

\item{\code{cv_mean_score}}{A data.frame of mean scores from cross-validation.}

\item{\code{test_metrics}}{A data.frame containing metrics (F1, accuracy, precision, and recall) calculated on the unseen test set. Contains mean values across splits as well as standard deviation.}
}}

