% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_importance.R
\name{plot_feature_importance}
\alias{plot_feature_importance}
\title{Plot Feature Importance}
\usage{
plot_feature_importance(pipelineresults, top_n_features = 10)
}
\arguments{
\item{pipelineresults}{An object of class \code{PipelineResults}.}

\item{top_n_features}{An integer specifying the top N features to plot based on their mean importance.}
}
\value{
A list of grid of ggplot objects.
}
\description{
This function plots the feature importance scores from \code{inbuilt_feature_importance} and \code{permutation_importance} in the \code{PipelineResults} object.
}
