% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{set_cache_options}
\alias{set_cache_options}
\title{Set Cache Options}
\usage{
set_cache_options(verbose = TRUE, max_cache_size = NULL)
}
\arguments{
\item{verbose}{Logical, print cache messages}

\item{max_cache_size}{Numeric, max cache size in MB (not implemented yet)}
}
\description{
Configure caching behavior
}
