% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{set_cache_options}
\alias{set_cache_options}
\title{Set Cache Options}
\usage{
set_cache_options(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical, print cache messages}
}
\value{
\code{NULL}, invisibly.
}
\description{
Configure caching behavior
}
\examples{
set_cache_options(verbose = FALSE)

}
