% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSelectR.R
\name{set_default_fs_methods}
\alias{set_default_fs_methods}
\title{Set Default Feature Selection Methods}
\usage{
set_default_fs_methods(modules, max_features, random_state)
}
\arguments{
\item{modules}{A list containing the definitions for the Python modules and submodules.}

\item{max_features}{The maximum number of features to consider.}
}
\value{
A list containing preprocessing steps and default feature selection methods.
}
\description{
Set Default Feature Selection Methods
}
