% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSelectR.R
\name{set_default_param_grids}
\alias{set_default_param_grids}
\title{Set Default Parameter Grids for Feature Selection}
\usage{
set_default_param_grids()
}
\value{
A list containing the default parameter grids for feature selection methods.
}
\description{
Set Default Parameter Grids for Feature Selection
}
