% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_check.R
\name{skip_if_no_modules}
\alias{skip_if_no_modules}
\title{Check if Python Modules are Available}
\usage{
skip_if_no_modules(module_names)
}
\arguments{
\item{module_names}{A vector of names of the Python modules to check.}
}
\description{
This helper function checks if a list of Python modules are available. If any are not, it skips the tests.
}
