% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSelectR.R
\name{split_data}
\alias{split_data}
\title{Split Data into Training and Test Sets}
\usage{
split_data(X, y, test_size, modules)
}
\arguments{
\item{X}{A dataframe or matrix of predictors.}

\item{y}{A vector of outcomes.}

\item{test_size}{Proportion of the data to be used as the test set.}

\item{modules}{A list containing the definitions for the Python modules and submodules.}
}
\value{
A list containing the training and test sets for predictors and outcomes.
}
\description{
Split Data into Training and Test Sets
}
