% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate GeneSelectR Inputs}
\usage{
validate_inputs(X, y, gene_names = NULL, n_folds = 5)
}
\arguments{
\item{X}{Expression matrix}

\item{y}{Outcome vector}

\item{gene_names}{Gene names}

\item{n_folds}{Number of CV folds}
}
\description{
Validate GeneSelectR Inputs
}
\keyword{internal}
